requirements.txt
elasticsearch>=8.0.0
pytest>=7.0.0
pillow>=9.0.0
numpy>=1.20.0
pandas>=1.3.0
fastapi>=0.95.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.28.0
pydantic>=1.10.0
httpx>=0.23.0
python-multipart>=0.0.5
openai>=1.0.0
